Refactor mask generation by enhancing mask preparation and validation processes
